moved socket io messaging into user profile page
